# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {
      "name": "API Support",
      "url": "https://github.com/fransfilastap/kontak"
    },
    "title": "Kontak API",
    "version": "1.0",
    "description": "WhatsApp-to-REST-API bridge",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "paths": {
    "/admin/broadcasts": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Get all broadcast jobs for the authenticated user",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "broadcasts"
        ],
        "summary": "List broadcasts",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "additionalProperties": true
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Create a new broadcast job to send messages to multiple recipients",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "broadcasts"
        ],
        "summary": "Create broadcast",
        "parameters": [
          {
            "description": "Broadcast data",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pkg_http.CreateBroadcastRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          }
        }
      }
    },
    "/admin/broadcasts/{id}": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Get a specific broadcast job with its recipients",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "broadcasts"
        ],
        "summary": "Get broadcast",
        "parameters": [
          {
            "type": "string",
            "description": "Broadcast Job ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          }
        }
      }
    },
...
    "/admin/contacts/{client_id}": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Get all contacts for a device",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "contacts"
        ],
        "summary": "List contacts",
        "parameters": [
          {
            "type": "string",
            "description": "Device ID",
            "name": "client_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "additionalProperties": true
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          }
        }
      }
    },
    "/admin/contacts/{client_id}/sync": {
      "put": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Fetch and sync contacts from WhatsApp for a device",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "contacts"
        ],
        "summary": "Sync contacts",
        "parameters": [
          {
            "type": "string",
            "description": "Device ID",
            "name": "client_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/pkg_http.ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "pkg_http.CreateBroadcastRequest": {
      "type": "object",
      "required": [
        "content",
        "device_id",
        "message_type",
        "name",
        "recipients"
      ],
      "properties": {
        "content": {
          "type": "string"
        },
        "cooldown": {
          "type": "integer"
        },
        "device_id": {
          "type": "string"
        },
        "message_type": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "recipients": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "scheduled_at": {
          "type": "string"
        }
      }
    },
...
    "pkg_http.ErrorResponse": {
      "type": "object",
      "properties": {
        "context": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
  },
  "host": "localhost:8080",
  "basePath": "/"
}